LLAMA-6895 : Change kwd sensitivity to a floating point
Reason for change: Change keyword sensitivity value from detector
specific integer values and ranges to flexible and consistent floating
point values. Limit checking to be handled by the detector.
Test Procedure: Test and verify sensitivity values in configuration
files are applied correctly to applicable keyword detectors
Risks: Low

(cherry picked from commit 841efb9db21464a5f60cb8a6d9ac8d0392c57a63)

+++ src/xrsv_ws/xrsv_ws.h
@@ -95,11 +95,11 @@ typedef struct {
    uint32_t keyword_sample_begin;               ///< The offset in samples from the beginning of the buffered audio to the beginning of the keyword
    uint32_t keyword_sample_end;                 ///< The offset in samples from the beginning of the buffered audio to the end of the keyword
    uint16_t keyword_doa;                        ///< The direction of arrival in degrees (0-359)
-   uint16_t keyword_sensitivity;                ///<
+   double   keyword_sensitivity;                ///<
    bool     keyword_sensitivity_triggered;      ///<
    bool     keyword_sensitivity_high_support;   ///<
    bool     keyword_sensitivity_high_triggered; ///<
-   uint16_t keyword_sensitivity_high;           ///<
+   double   keyword_sensitivity_high;           ///<
    double   keyword_gain;                       ///< The gain applied to the audio at the input of the keyword detector
    double   dynamic_gain;                       ///< The gain applied to the streamed audio
    double   signal_noise_ratio;                 ///<
